requirements.txt
Pillow
PyQt5
pyinstaller